Update README.md
I tried my best to "replicate" the page. While it is possible to just copy each element from the website along with the styles, that becomes redundant and tedious. With enough time I could make it nearly identical. I think the main test was to see if I could implement the jQuery table by following that sites instructions, which was straight forward.